backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.125.0",
    "firebase-admin>=6.6.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.14.10",
    "uvicorn>=0.38.0",
]
